+++ src/commands/task/delete.ts
