+++ src/commands/task/groups/list.ts
@@ -2,9 +2,9 @@
 
 import { CliUx } from '@oclif/core'
 
-import { Command } from '../../lib/command'
-import * as flags from '../../lib/flags'
-import {  printTaskGroups } from '../../lib/utils'
+import { Command } from '../../../lib/command'
+import * as flags from '../../../lib/flags'
+import {  printTaskGroups } from '../../../lib/utils'
 // eslint-disable-next-line quotes
 import debugFn = require('debug')
 import { isEmpty } from 'lodash'
