+++ src/commands/task/start.ts
@@ -38,6 +38,13 @@ export default class TasksStartCommand extends Command {
 
     let encoded_string = startArgs.args as string
     if (encoded_string.charAt(0) == `@`) {
+
+      const stats = fs.statSync(encoded_string.substring(1, encoded_string.length))
+      const fileSizeInMegabytes = stats.size / (1024*1024)
+      if (fileSizeInMegabytes > 10) {
+        this.error(`File is too large`)
+      }
+
       encoded_string = fs.readFileSync(encoded_string.substring(1, encoded_string.length),{ encoding: `utf8`, flag: `r` }).toString()
     }
