+++ src/commands/task/types/list/majors.ts
@@ -2,16 +2,16 @@
 
 import { CliUx } from '@oclif/core'
 
-import { Command } from '../../../lib/command'
-import * as flags from '../../../lib/flags'
-import { printMinors } from '../../../lib/utils'
+import { Command } from '../../../../lib/command'
+import * as flags from '../../../../lib/flags'
+import {  printMajors } from '../../../../lib/utils'
 // eslint-disable-next-line quotes
 import debugFn = require('debug')
 import { isEmpty } from 'lodash'
 
 const debug = debugFn(`task-types:list`)
 
-export default class TaskTypesListMinorsCommand extends Command {
+export default class TaskTypesListMajorsCommand extends Command {
   static description = `List task type configurations`
   static strict = false
 
@@ -19,7 +19,7 @@ export default class TaskTypesListMinorsCommand extends Command {
 
   static flags = {
     ...flags.subscriber,
-    ...CliUx.ux.table.flags()
+    ...CliUx.ux.table.flags(),
   }
 
   static args = [
@@ -33,31 +33,24 @@ export default class TaskTypesListMinorsCommand extends Command {
       name: `type`,
       required: true,
       description: `Task type name`,
-    },
-    {
-      name: `major`,
-      required: true,
-      description: `Major version`,
     }
   ]
-
   async run(): Promise<void> {
-    const { flags, argv } = await this.parse(TaskTypesListMinorsCommand)
+    const { flags, argv } = await this.parse(TaskTypesListMajorsCommand)
     const subscriberId = flags[`subscriber-id`]
     const namespace = argv[0] as string
     const type = argv[1] as string
-    const major = argv[2] as string
 
     try {
-      const minors = await this.relay.fetchMinors(subscriberId, namespace, type, major)
+      const majors = await this.relay.fetchMajors(subscriberId, namespace, type)
 
-      debug(`minors`, minors)
+      debug(`majors`, majors)
 
-      if (isEmpty(minors)) {
-        this.error(`No minors found. Check namespace, type and major args.`)
+      if (isEmpty(majors)) {
+        this.error(`No majors found: Check namespace and type args.`)
       }
 
-      printMinors(minors, flags, type, false, namespace)
+      printMajors(majors, flags, type, namespace)
 
     } catch (err) {
       debug(err)
@@ -66,6 +59,3 @@ export default class TaskTypesListMinorsCommand extends Command {
 
   }
 }
-
-
-
