+++ src/commands/task/types/update/minor.ts
@@ -1,28 +1,28 @@
 // Copyright © 2023 Relay Inc.
 
-import * as flags from '../../../lib/flags'
+import * as flags from '../../../../lib/flags'
 // eslint-disable-next-line quotes
 import debugFn = require('debug')
 
-import { updateMajor } from '../../../lib/task-types'
-import { NewMajor } from '../../../lib/api'
-import { Command } from '../../../lib/command'
+import { updateMinor } from '../../../../lib/task-types'
+import { NewMinor } from '../../../../lib/api'
+import { Command } from '../../../../lib/command'
 import { readFileSync } from 'fs'
 
 import util from 'node:util'
 // eslint-disable-next-line @typescript-eslint/no-var-requires
 const exec = util.promisify(require(`node:child_process`).exec)
 // eslint-disable-next-line @typescript-eslint/no-var-requires
-const debug = debugFn(`task-types:update:major`)
+const debug = debugFn(`task-types:update:minor`)
 
 async function execute (command: string): Promise<string> {
   const { stdout } = (await exec(command))
   return stdout.toString().trimEnd()
 }
 
-export default class TaskTypesUpdateMajorCommand extends Command {
+export default class TaskTypesUpdateMinorCommand extends Command {
 
-  static description = `Update a task type's major. Must have admin priviledges and RELAY_ADMIN_TOKEN env variable set to run this command.`
+  static description = `Update a task type. Must have admin priviledges and RELAY_ADMIN_TOKEN env variable set to run this command.`
   static strict = false
 
   // static hidden = true
@@ -35,7 +35,7 @@ export default class TaskTypesUpdateMajorCommand extends Command {
       multiple: false,
       description: `Git version of source file`,
       helpValue: `<branch>@<commit hash>`
-    }),
+    })
   }
 
   static args = [
@@ -51,18 +51,24 @@ export default class TaskTypesUpdateMajorCommand extends Command {
       description: `Task type name`,
     },
     {
+      name: `major`,
+      required: true,
+      description: `Major version`,
+    },
+    {
       name: `source`,
       required: true,
       description: `Capsule source file name`,
-    }
+    },
   ]
 
   async run(): Promise<void> {
-    const { flags, argv } = await this.parse(TaskTypesUpdateMajorCommand)
+    const { flags, argv } = await this.parse(TaskTypesUpdateMinorCommand)
     const subscriberId = flags[`subscriber-id`]
     const namespace = argv[0] as string
     const name = argv[1] as string
-    let source = argv[2] as string
+    const major = argv[2] as string
+    let source = argv[3] as string
     try {
       if (!flags.key) {
         const scriptDir = (await execute(`dirname ${source}`))
@@ -70,14 +76,13 @@ export default class TaskTypesUpdateMajorCommand extends Command {
         const gitCommit = (await execute(`cd ${scriptDir} && git rev-parse HEAD`))
         flags.key = `${gitBranch}@${gitCommit}`
       }
-
       source = readFileSync(source, `utf-8`)
-      const major: NewMajor = await updateMajor(flags, source) as NewMajor
-      const success = await this.relay.createMajor(subscriberId, name, namespace, major)
+      const minor: NewMinor = await updateMinor(flags, source) as NewMinor
+      const success = await this.relay.createMinor(subscriberId, name, namespace, major, minor)
       if (success) {
-        this.log(`Successfully updated major`)
+        this.log(`Successfully updated minor`)
       } else {
-        this.log(`Could not update major`)
+        this.log(`Could not update minor`)
       }
     } catch (err) {
       debug(err)
